Layout example app for gridview
