--- src/_includes/code/layout/grid/main.dart
@@ -1,68 +1,0 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'package:flutter/material.dart';
-// Uncomment lines 7 and 10 to view the visual layout at runtime.
-//import 'package:flutter/rendering.dart' show debugPaintSizeEnabled;
-
-void main() {
-  //debugPaintSizeEnabled = true;
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-// The images are saved with names pic1.jpg, pic2.jpg...pic30.jpg.
-// The List.generate constructor allows an easy way to create
-// a list when objects have a predictable naming pattern.
-List<Container> _buildGridTileList(int count) {
-  List<Container> containers = List<Container>.generate(
-      count,
-      (int index) =>
-          Container(child: Image.asset('images/pic${index+1}.jpg')));
-  return containers;
-}
-
-Widget buildGrid() {
-  return GridView.extent(
-      maxCrossAxisExtent: 150.0,
-      padding: const EdgeInsets.all(4.0),
-      mainAxisSpacing: 4.0,
-      crossAxisSpacing: 4.0,
-      children: _buildGridTileList(30));
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: buildGrid(),
-      ),
-    );
-  }
-}
