--- src/_includes/code/layout/grid/pubspec.yaml
@@ -1,40 +1,0 @@
-name: grid
-description: Demonstrates GridView.extent.
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
-  assets:
-    - images/pic1.jpg
-    - images/pic2.jpg
-    - images/pic3.jpg
-    - images/pic4.jpg
-    - images/pic5.jpg
-    - images/pic6.jpg
-    - images/pic7.jpg
-    - images/pic8.jpg
-    - images/pic9.jpg
-    - images/pic10.jpg
-    - images/pic11.jpg
-    - images/pic12.jpg
-    - images/pic13.jpg
-    - images/pic14.jpg
-    - images/pic15.jpg
-    - images/pic16.jpg
-    - images/pic17.jpg
-    - images/pic18.jpg
-    - images/pic19.jpg
-    - images/pic20.jpg
-    - images/pic21.jpg
-    - images/pic22.jpg
-    - images/pic23.jpg
-    - images/pic24.jpg
-    - images/pic25.jpg
-    - images/pic26.jpg
-    - images/pic27.jpg
-    - images/pic28.jpg
-    - images/pic29.jpg
-    - images/pic30.jpg
